Envio de sinais funcionando

+++ Trabalho 1/sinais_sender.cpp
@@ -8,8 +8,7 @@ using namespace std;
 
 
 int main (int argc, char *argv[]) {
-   // register signal SIGINT and signal handler 
-
+   
    if(argc<=1) {
         cout<<"You did not feed me arguments, I will die now :( ..."<<endl;
         exit(1);
@@ -18,12 +17,19 @@ int main (int argc, char *argv[]) {
    pid_t pidToSend = atoi(argv[1]);
    int sig = atoi(argv[2]);
 
-   //TODO: dar erro quando processo não existir
-   cout<< "Sending signal " << sig << " to process " << pidToSend << endl;
+   if (kill(pidToSend, 0)==0){ 
 
-   kill(pidToSend, sig);
+      cout<< "Enviando sinal " << sig << " para o processo " << pidToSend << endl;
 
+      kill(pidToSend, sig);
 
+   }
+
+   else{
+
+      cout<< "Erro - O processo " << pidToSend << " não existe" << endl;
+
+   }
 
    return 0;
 }
